% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dohactheme.R
\docType{data}
\name{dohac.blue}
\alias{dohac.blue}
\title{DoHAC Blue}
\format{
\subsection{\code{dohac.blue}}{

An object of class \code{colour}.
}
}
\source{
\url{https://github.com/zerogetsamgow/dohactheme}
}
\usage{
dohac.blue
}
\description{
Blue is used in majority
}
\examples{
plot(dohac.blue)
}
\seealso{
\code{\link[=pal_colour]{pal_colour()}}
}
\keyword{datasets}
