% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dohactheme.R
\docType{data}
\name{dohac.teal}
\alias{dohac.teal}
\title{DoHAC Teal}
\format{
\subsection{\code{dohac.teal}}{

An object of class \code{colour}.
}
}
\usage{
dohac.teal
}
\description{
Teal is the secondary colour.
}
\examples{
plot(dohac.teal)
}
\keyword{datasets}
