% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dohactheme.R
\docType{data}
\name{dohac_colours}
\alias{dohac_colours}
\title{Cube Colours}
\format{
\subsection{\code{dohace_colours}}{

An object of class \code{palette} with 5 colours.
Use \code{names(dohac_colours)} to return all colour names.
}
}
\source{
\url{https://github.com/zerogetsamgow/dohactheme}
}
\usage{
dohac_colours
}
\description{
Department of Health and Aged Care's colours play an important role in the
Department’s identity. The colours included in this section are able to be
used across various media.
}
\examples{
# Plot all colours
plot(dohac_colours)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}#'
}
\keyword{datasets}
