% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dohactheme-package.R
\docType{package}
\name{dohactheme-package}
\alias{dohactheme}
\alias{dohactheme-package}
\title{dohactheme: DoHAC Themes Colour Palettes}
\description{
A colour palette and ggplot2 themes that comply with the AustralianDepartment of Health and Aged Care Sytle SheetBuilt using the 'palettes' package, which provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, colour mixing arithmetic, and colour interpolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/user/dohactheme}
  \item Report bugs at \url{https://github.com/user/dohactheme/issues}
}

}
\author{
\strong{Maintainer}: Samuel Gow \email{samuel.gow@cubegroup.com.au}

}
\keyword{internal}
