% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dohac_fonts.R
\name{install_dohac_fonts}
\alias{install_dohac_fonts}
\title{Install Cube Fonts}
\usage{
install_dohac_fonts(font.path = "C:/windows/fonts")
}
\description{
Themes in this package use the Agenda font.
}
\details{
This function allows these to be installed within the package
}
\concept{cubethemes}
