% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dohac_blue.R
\name{theme_dohac_blue}
\alias{theme_dohac_blue}
\title{DoHAC Blue Theme}
\usage{
theme_dohac_blue(
  base.colour = dohac.blue,
  base.size = 15,
  line.colour = "white",
  text.colour = "white"
)
}
\description{
This theme takes \code{\link[dohactheme]{theme_cube_foundations}()} and sets
background colour to blue and line and text colours to white.
}
\details{
Colours are sourced from \code{dohactheme} and are DoHAC style guide compliant.
}
\seealso{
Other dohactheme: 
\code{\link{theme_dohac_foundation}()},
\code{\link{theme_dohac_white}()}
}
\concept{dohactheme}
