% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dohac_foundation.R
\name{theme_dohac_foundation}
\alias{theme_dohac_foundation}
\title{DoHAC Foundation Theme}
\usage{
theme_dohac_foundation(base_size = 15, base_family = "Arial")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This theme is designed to be a foundation from which to build new
themes, and not meant to be used directly. \code{theme_dohac_foundation()}
is a complete theme with only minimal number of elements defined.
It is easier to create new Cube themes by extending this one rather
than \code{\link[ggplot2]{theme_gray}()} or \code{\link[ggplot2]{theme_bw}()},
because those themes define elements deep in the hierarchy.
}
\details{
This theme takes \code{\link[ggthemes]{theme_foundations}()} and sets
base text family and size to \verb{Helvetica neue} and \code{15}.
}
\seealso{
Other dohactheme: 
\code{\link{theme_dohac_blue}()},
\code{\link{theme_dohac_white}()}
}
\concept{dohactheme}
