% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dohac_map.R
\name{theme_dohac_map}
\alias{theme_dohac_map}
\title{DoHAC Map Theme}
\usage{
theme_dohac_map(
  base_colour = c("white", "blue"),
  base_size = 15,
  base_family = "Helvetica neue"
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This theme is designed to be be used when creating maps.
}
\details{
This theme takes \code{\link[ggthemes]{theme_map}()} and sets
base text family and size to \verb{Helvetica neue} and \code{15}.
}
\concept{dohacthemes}
