% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dohac_white.R
\name{theme_dohac_white}
\alias{theme_dohac_white}
\title{DoHAC White Theme}
\usage{
theme_dohac_white(
  base.colour = "white",
  line.colour = dohac.blue,
  text.colour = dohac.blue
)
}
\description{
This theme takes \code{\link[dohactheme]{theme_cube_foundations}()} and sets
background colour to white and line and text colours to blue.
}
\details{
Colours are sourced from \code{dohactheme} and are DoHAC style guide compliant.
}
\seealso{
Other dohactheme: 
\code{\link{theme_dohac_blue}()},
\code{\link{theme_dohac_foundation}()}
}
\concept{dohactheme}
